backend/app/utils/auth.py
from datetime import datetime, timedelta
from jose import JWTError, jwt
from app.config import settings
from app.schemas.user import UserCreate

# Create a JWT token
def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(hours=1)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.JWT_SECRET_KEY, algorithm=settings.JWT_ALGORITHM)
    return encoded_jwt
